# openapi.yaml
openapi: 3.0.3
info:
  title: Exam Integrity Engine API
  description: |
    Production API for the Exam Integrity Engine -- a Vietnamese-education exam digitisation,
    proctoring, and scoring platform.

    **Base path**: `/exam-integrity-backend`
    **Authentication**: All endpoints require `Authorization: Bearer <keycloak-jwt>` (OAuth2 Resource Server).

    ## Workflow Overview

    ### Teacher (Draft) Workflow
    ```
    POST /api/drafts (upload PDF)
      -> GET  /api/drafts/{draftId}                        (open for review, locks draft)
      -> PATCH /api/drafts/{draftId}/questions/{questionId} (correct questions)
      -> DELETE /api/drafts/{draftId}/questions/{questionId} (remove question, re-sequence)
      -> GET  /api/questions                               (browse bank for a replacement)
      -> POST /api/drafts/{draftId}/questions              (add bank question or inline)
      -> POST /api/drafts/{draftId}/publish               (create live Exam)
    ```

    ### Student (Session) Workflow
    ```
    POST /api/sessions                               (create session, starts Redis timer)
      -> GET  /api/sessions/{sessionId}/questions/N  (load each question)
      -> PATCH /api/sessions/{sessionId}/answers/{questionId}  (auto-save answers)
      -> POST /api/sessions/{sessionId}/submit       (submit, triggers async scoring)
      -> GET  /api/sessions/{sessionId}/review       (poll until scoring complete)
    ```

...
  version: 1.0.0
  contact:
...
servers:
...
paths:
...
  /api/sessions:
    post:
      tags: [Exam Session]
      operationId: createSession
      summary: Create a new exam session
      description: |
        Starts a new exam session for a student.

        - Initialises Redis timer key `timer:{sessionId}` with TTL = `exam.durationSeconds`
        - Returns `remainingSeconds` for the UI's initial countdown render
        - Publishes `exam.ingested` Kafka event for audit
      parameters:
        - name: examId
          in: query
          required: true
          description: ID of the published exam
          example: exam-uuid
          schema:
            type: string
        - name: studentId
          in: query
          required: true
          description: Keycloak subject (sub claim) of the student
          example: student-uuid
          schema:
            type: string
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDTO'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Student already has an ACTIVE session for this exam
  /api/sessions/{sessionId}:
    get:
      tags: [Exam Session]
      operationId: getSession
      summary: Get current session state
      description: Returns `status` and `remainingSeconds` (read from Redis).
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDTO'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/sessions/{sessionId}/timer:
    get:
      tags: [Exam Session]
      operationId: getTimer
      summary: Get remaining seconds (REST fallback)
      description: |
        REST fallback for clients that cannot use WebSocket.
        Returns plain `long` value.

        **Prefer** STOMP subscription to `/topic/session/{sessionId}` for real-time 1-second ticks.
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Remaining seconds
          content:
            application/json:
              schema:
                type: integer
                format: int64
                example: 2513
  /api/sessions/{sessionId}/questions/{questionNumber}:
    get:
      tags: [Exam Session]
      operationId: getQuestion
      summary: Get one question for display
      description: |
        Returns question content for the student to answer.
        `questionNumber` is **1-based** (first question = 1).

        **Security**: `correctAnswer` and `rubric` are **never** included in this response.
        They are only revealed in the ReviewDashboard after submission.
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - name: questionNumber
          in: path
          required: true
          description: 1-based question index
          example: 3
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Question for display
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionSummaryDTO'
        '403':
          description: Session is not ACTIVE
        '404':
          $ref: '#/components/responses/NotFound'
  /api/sessions/{sessionId}/answers/{questionId}:
    patch:
      tags: [Exam Session]
      operationId: saveAnswer
      summary: Save or update an answer
      description: |
        Saves the student's answer for one question.
        Idempotent -- last write wins. Updates `ExamSession.answers[questionId]` in MongoDB.

        **When to call**: automatically on every page navigation (Next / Previous button)
        and whenever the student toggles the "Flag for review" checkbox.
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - $ref: '#/components/parameters/questionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerDTO'
            example:
              questionId: q-uuid-003
              answer: "A"
              flaggedForReview: false
      responses:
        '204':
          description: Saved
        '403':
          description: Session is not ACTIVE -- cannot modify answers after submission
        '404':
          $ref: '#/components/responses/NotFound'
  /api/sessions/{sessionId}/submit:
    post:
      tags: [Exam Session]
      operationId: submitExam
      summary: Submit the exam
      description: |
        Marks the session as `SUBMITTED` and publishes an `exam.submitted` Kafka event.
        Scoring runs **asynchronously** via the scoring-worker service.

        **UI flow**:
        1. Show confirmation modal: "Are you sure? You cannot change your answers after submission."
        2. On confirm, call this endpoint.
        3. On `202`, redirect student to `/review/{sessionId}` and begin polling.
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '202':
          description: Accepted -- scoring started asynchronously
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Session already submitted or force-submitted
  /api/sessions/{sessionId}/review:
    get:
      tags: [Exam Session]
      operationId: getReviewDashboard
      summary: Get review dashboard (poll until complete)
      description: |
        Returns scores and corrections for all questions after submission.

        **Polling**: Poll every 2 seconds. Scoring is complete when no `ScoreDTO`
        has `status=PENDING_ESSAY`.

        **Score formula**:
        - MCQ: `CORRECT` (full points) or `INCORRECT` (zero)
        - Essay: `EssayScore = MaxPoint x (0.35xKeyword + 0.25xSteps + 0.20xFinalAnswer + 0.15xSemantic + 0.05xFormat)`
        - Final: `finalScore10 = (totalEarned / totalMax) x 10`
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Review dashboard ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDashboardDTO'
        '202':
          description: Scoring still in progress -- keep polling
        '404':
          $ref: '#/components/responses/NotFound'

  # -- Proctoring -----------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT. Configure JWK set URI via KEYCLOAK_JWK_SET_URI.
  parameters:
...
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      example: session-uuid
    questionId:
      name: questionId
      in: path
      required: true
      schema:
        type: string
      example: q-uuid-003
  responses:
    NotFound:
      description: Resource not found
  schemas:
...
    QuestionSummaryDTO:
      type: object
      description: Question for student display. correctAnswer and rubric are never included.
      properties:
        id:
          type: string
          example: q-uuid-003
        questionNumber:
          type: integer
          example: 3
        content:
          type: string
          example: "Tinh 12 x 34 = ?"
        type:
          type: string
          enum: [MCQ, ESSAY_SHORT, ESSAY_LONG]
        points:
          type: number
          format: double
          example: 0.5
        options:
          type: array
          nullable: true
          description: Present only for MCQ questions
          items:
            type: string
          example: ["A. 408", "B. 418", "C. 480", "D. 488"]
        truncated:
          type: boolean
          description: True if OCR cut off part of this question
          example: false

    # -- Draft ----------------------------------------------------------------
...
    SessionDTO:
      type: object
      properties:
        sessionId:
          type: string
          example: session-uuid
        examId:
          type: string
          example: exam-uuid
        studentId:
          type: string
          example: student-uuid
        status:
          type: string
          enum: [ACTIVE, SUBMITTED, FORCE_SUBMITTED, FLAGGED]
          example: ACTIVE
        startedAt:
          type: string
          format: date-time
          example: "2024-01-15T09:00:00Z"
        remainingSeconds:
          type: integer
          format: int64
          example: 2513
    AnswerDTO:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: string
          example: q-uuid-003
        answer:
          type: string
          description: MCQ choice letter (A/B/C/D) or essay free-text
          example: "A"
        flaggedForReview:
          type: boolean
          default: false
          description: True if student marked this question for review before submitting
          example: false

    # -- Review / Scoring -----------------------------------------------------
    ReviewDashboardDTO:
      type: object
      properties:
        sessionId:
          type: string
          example: session-uuid
        totalEarned:
          type: number
          format: double
          example: 8.5
        totalMax:
          type: number
          format: double
          example: 10.0
        finalScore10:
          type: number
          format: double
          description: "Final score on a 10-point scale: (totalEarned / totalMax) x 10"
          example: 8.5
        missedQuestionNumbers:
          type: array
          items:
            type: integer
          description: 1-based question numbers where status != CORRECT
          example: [3, 11, 17]
        scores:
          type: array
          items:
            $ref: '#/components/schemas/ScoreDTO'
    ScoreDTO:
      type: object
      properties:
        questionId:
          type: string
          example: q-uuid-003
        questionNumber:
          type: integer
          example: 3
        earnedPoints:
          type: number
          format: double
          example: 0.4
        maxPoints:
          type: number
          format: double
          example: 0.5
        status:
          type: string
          enum:
            - CORRECT
            - INCORRECT
            - PARTIAL
            - PENDING_ESSAY
            - INCOMPLETE_QUESTION
            - MULTIPLE_ANSWERS_FLAG
            - SELF_GRADE_REQUIRED
          example: PARTIAL
        studentAnswer:
          type: string
          example: "Paris"
        correctAnswer:
          type: string
          description: Only revealed here after submission
          example: "Paris"
        explanation:
          type: string
          example: "Keyword 'buoc giai' missing from answer"
        scoreBreakdown:
          $ref: '#/components/schemas/ScoreBreakdownDTO'
    ScoreBreakdownDTO:
      type: object
      nullable: true
      description: Detailed scoring breakdown. Non-null only for ESSAY_SHORT and ESSAY_LONG questions.
      properties:
        keywordScore:
          type: number
          format: double
          description: Keyword match ratio (weight 35%)
          example: 0.90
        stepsScore:
          type: number
          format: double
          description: Reasoning steps match ratio (weight 25%)
          example: 0.80
        finalAnswerScore:
          type: number
          format: double
          description: Final answer correctness (weight 20%)
          example: 1.0
        semanticScore:
          type: number
          format: double
          description: Vietnamese SBERT semantic similarity (weight 15%)
          example: 0.85
        formatScore:
          type: number
          format: double
          description: Format / unit requirement score (weight 5%)
          example: 0.70
        weightedTotal:
          type: number
          format: double
          description: "Weighted sum: 0.35*kw + 0.25*steps + 0.20*fa + 0.15*sem + 0.05*fmt"
          example: 0.87
security:
  - bearerAuth: []